Move up tasks subdir
