+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "heareval/tasks/secrettasks"]
-	path = hearpreprocess/tasks/secrettasks
+	path = hearpreprocess/secrettasks
 	url = git@github.com:neuralaudio/hear2021-secret-tasks
