+++ hearpreprocess/office_events.py
@@ -4,7 +4,9 @@ Pre-processing pipeline for DCASE 2016 task 2 task (sound event
 detection).
 
 The HEAR 2021 variation of DCASE 2016 Task 2 is that we ignore the
-monophonic training data and use the dev data for train.
+monophonic training data. We mix the dev and eval data, and
+re-partition ourselves, to reduce variance between the partitions.
+
 We also allow training data outside this task.
 """
 
@@ -12,7 +14,7 @@ import logging
 from pathlib import Path
 from typing import List
 
-import hearpreprocess.tasks.pipeline as pipeline
+import hearpreprocess.pipeline as pipeline
 import luigi
 import pandas as pd
 from slugify import slugify
@@ -20,18 +22,20 @@ from slugify import slugify
 logger = logging.getLogger("luigi-interface")
 
 task_config = {
-    "task_name": "dcase2016_task2",
+    "task_name": "office_events",
     "version": "hear2021",
     "embedding_type": "event",
     "prediction_type": "multilabel",
     "download_urls": [
         {
             "split": "train",
+            "name": "dev",
             "url": "https://archive.org/download/dcase2016_task2_train_dev/dcase2016_task2_train_dev.zip",  # noqa: E501
             "md5": "4e1b5e8887159193e8624dec801eb9e7",
         },
         {
-            "split": "test",
+            "split": "train",
+            "name": "eval",
             "url": "https://archive.org/download/dcase2016_task2_test_public/dcase2016_task2_test_public.zip",  # noqa: E501
             "md5": "ac98768b39a08fc0c6c2ddd15a981dd7",
         },
@@ -41,11 +45,13 @@ task_config = {
         "download_urls": [
             {
                 "split": "train",
+                "name": "dev",
                 "url": "https://github.com/neuralaudio/hear2021-open-tasks-downsampled/raw/main/dcase2016_task2_train_dev-small.zip",  # noqa: E501
                 "md5": "aa9b43c40e9d496163caab83becf972e",
             },
             {
-                "split": "test",
+                "split": "train",
+                "name": "eval",
                 "url": "https://github.com/neuralaudio/hear2021-open-tasks-downsampled/raw/main/dcase2016_task2_test_public-small.zip",  # noqa: E501
                 "md5": "14539d85dec03cb7ac75eb62dd1dd21e",
             },
@@ -61,11 +67,11 @@ task_config = {
 
 
 class ExtractMetadata(pipeline.ExtractMetadata):
-    train = luigi.TaskParameter()
-    test = luigi.TaskParameter()
+    train_dev = luigi.TaskParameter()
+    train_eval = luigi.TaskParameter()
 
     def requires(self):
-        return {"train": self.train, "test": self.test}
+        return {"train_eval": self.train_eval, "train_dev": self.train_dev}
 
     @staticmethod
     def slugify_file_name(relative_path: str):
@@ -84,27 +90,28 @@ class ExtractMetadata(pipeline.ExtractMetadata):
     """
     DCASE 2016 uses funny pathing, so we just hardcode the desired
     (paths)
-    Note that for our training data, we only use DCASE 2016 dev data.
-    Their training data is short monophonic events.
     """
-    split_to_path_str = {
-        "train": "dcase2016_task2_train_dev/dcase2016_task2_dev/",
-        "test": "dcase2016_task2_test_public/",
+    requires_key_to_path_str = {
+        "train_dev": "train/dev/dcase2016_task2_train_dev/dcase2016_task2_dev/",
+        "train_eval": "train/eval/dcase2016_task2_test_public/",
     }
 
-    def get_requires_metadata(self, split: str) -> pd.DataFrame:
-        logger.info(f"Preparing metadata for {split}")
+    def get_requires_metadata(self, requires_key: str) -> pd.DataFrame:
+        logger.info(f"Preparing metadata for {requires_key}")
 
-        split_path = (
-            Path(self.requires()[split].workdir)
-            .joinpath(split)
-            .joinpath(self.split_to_path_str[split])
+        assert requires_key.startswith("train_")
+
+        requires_path = Path(self.requires()[requires_key].workdir).joinpath(
+            self.requires_key_to_path_str[requires_key]
         )
 
         metadatas = []
-        for annotation_file in split_path.glob("annotation/*.txt"):
+        for annotation_file in requires_path.glob("annotation/*.txt"):
             metadata = pd.read_csv(
-                annotation_file, sep="\t", header=None, names=["start", "end", "label"]
+                annotation_file,
+                sep="\t",
+                header=None,
+                names=["start", "end", "label"],
             )
             # Convert start and end times to milliseconds
             metadata["start"] *= 1000
@@ -116,7 +123,7 @@ class ExtractMetadata(pipeline.ExtractMetadata):
             )
             metadata = metadata.assign(
                 relpath=sound_file,
-                split=lambda df: split,
+                split=lambda df: "train",
             )
 
             metadatas.append(metadata)
