+++ hearpreprocess/runner.py
@@ -8,17 +8,17 @@ import multiprocessing
 from typing import Optional
 
 import click
-import hearpreprocess.tasks.dcase2016_task2 as dcase2016_task2
-import hearpreprocess.tasks.nsynth_pitch as nsynth_pitch
-import hearpreprocess.tasks.office_events as office_events
-import hearpreprocess.tasks.pipeline as pipeline
-import hearpreprocess.tasks.speech_commands as speech_commands
+import hearpreprocess.dcase2016_task2 as dcase2016_task2
+import hearpreprocess.nsynth_pitch as nsynth_pitch
+import hearpreprocess.office_events as office_events
+import hearpreprocess.pipeline as pipeline
+import hearpreprocess.speech_commands as speech_commands
 
 logger = logging.getLogger("luigi-interface")
 # Currently the runner is only allowed to run for open tasks
 # The secret tasks module will be not be available for the participants
 try:
-    from hearpreprocess.tasks.secrettasks import hearsecrettasks
+    from hearpreprocess.secrettasks import hearsecrettasks
 
     secret_tasks = hearsecrettasks.tasks
